pyproject.toml
[project]
name = "inktime"
version = "0.1.0"
description = "E-ink photo frame system with VLM-powered photo analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "requests>=2.32.0",
    "pillow>=12.0.0",
    "pydantic-settings>=2.0.0",
]

[project.scripts]
photo_analyzer = "photo_analyzer.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["photo_analyzer"]

[dependency-groups]
dev = []
